requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==46.0.5
cloudinary==1.44.1
